changed user ID

+++ src/main/java/com/brumethon/app/domain/user/UserRepository.java
@@ -2,5 +2,5 @@ package com.brumethon.app.domain.user;
 
 import com.brumethon.kernel.Repository;
 
-public interface UserRepository extends Repository<User, Integer> {
+public interface UserRepository extends Repository<User, Long> {
 }
